refactor: auth Viewmodel 통합
