+++ Spender/app/src/main/java/com/example/spender/feature/auth/ui/KakaoLogin.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import com.example.spender.R
-import com.example.spender.feature.auth.ui.viewmodel.SocialViewModel
+import com.example.spender.feature.auth.ui.viewmodel.AuthViewModel
 import com.example.spender.ui.theme.BlackColor
 import com.example.spender.ui.theme.KakaoColor
 import com.example.spender.ui.theme.KakaoLabelColor
@@ -34,7 +34,7 @@ import com.example.spender.ui.theme.WhiteColor
 fun KakaoLogin(
     navController: NavHostController
 ) {
-    val viewModel: SocialViewModel = hiltViewModel()
+    val viewModel: AuthViewModel = hiltViewModel()
     val context = LocalContext.current
     Card(
         modifier = Modifier
