--- Spender/app/src/main/java/com/example/spender/feature/auth/ui/viewmodel/SocialViewModel.kt
@@ -1,102 +1,0 @@
-package com.example.spender.feature.auth.ui.viewmodel
-
-import android.app.Application
-import android.content.Context
-import android.util.Log
-import androidx.lifecycle.AndroidViewModel
-import androidx.lifecycle.viewModelScope
-import androidx.navigation.NavController
-import com.example.spender.core.data.remote.auth.LoginType
-import com.example.spender.feature.auth.data.AuthPrefs
-import com.example.spender.feature.auth.domain.AuthRepository
-import com.example.spender.feature.onboarding.data.OnboardingPref
-import com.google.firebase.Firebase
-import com.google.firebase.auth.auth
-import kotlinx.coroutines.launch
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class SocialViewModel @Inject constructor(
-    application: Application,
-    private val authRepository: AuthRepository
-) : AndroidViewModel(application) {
-
-    fun naverLogin(context: Context, navController: NavController) {
-        viewModelScope.launch {
-            try {
-                authRepository.naverLogin(context)
-                AuthPrefs.setLoginType(getApplication(), LoginType.NAVER)
-                Log.d("Login", "Naver Login Success!")
-                val currentUser = Firebase.auth.currentUser
-                val userMap: Map<String, Any?>? = currentUser?.let {
-                    mapOf(
-                        "uid" to it.uid,
-                        "email" to it.email,
-                        "name" to it.displayName,
-                        "photoUrl" to it.photoUrl?.toString()
-                    )
-                }
-
-                Log.d("Login", "naver User : ${userMap.toString()}")
-
-                navController.navigate(
-                    if (OnboardingPref.wasShown(getApplication())) "main"
-                    else "onboarding") {
-                    popUpTo("auth") { inclusive = true }
-                }
-            } catch (e: Exception) {
-                Log.e("Login", "Naver Login Fail", e)
-            }
-        }
-    }
-
-    fun kakaoLogin(context: Context, navController: NavController) {
-        viewModelScope.launch {
-            try {
-                authRepository.kakaoLogin(context)
-                AuthPrefs.setLoginType(getApplication(), LoginType.KAKAO)
-                Log.d("Login", "Kakao Login Success!")
-                val currentUser = Firebase.auth.currentUser
-                val userMap: Map<String, Any?>? = currentUser?.let {
-                    mapOf(
-                        "uid" to it.uid,
-                        "email" to it.email,
-                        "name" to it.displayName,
-                        "photoUrl" to it.photoUrl?.toString()
-                    )
-                }
-                Log.d("Login", "kakao User : ${userMap.toString()}")
-
-                navController.navigate(
-                    if (OnboardingPref.wasShown(getApplication())) "main"
-                    else "onboarding") {
-                    popUpTo("auth") { inclusive = true }
-                }
-            } catch (e: Exception) {
-                Log.e("Login", "Kakao Login Fail", e)
-            }
-        }
-    }
-
-    fun logout(context: Context, onDone: () -> Unit) {
-        viewModelScope.launch {
-            authRepository.logoutUser(context)
-            onDone()
-        }
-    }
-
-    fun withdraw(context: Context, onSuccess: () -> Unit, onError: (String) -> Unit) {
-        viewModelScope.launch {
-            try {
-                authRepository.withdrawUser(context)
-                onSuccess()
-            } catch (e: Exception) {
-                Log.e("Withdraw", "Unlink Failed", e)
-                onError(e.message ?: "error")
-            }
-        }
-    }
-
-
-}
